Code reorganization and imporoved configurability
Python data-generating code is now encompassed in a single src.py file that allows for single-click execution and easier maintainability.
Lots of functions have their prototypes documented and got commented extensively.
Lots of them got extented functionality to allow for external configuration.

config.py was better documented and extended to include optional specific data source options and so on.
src.py was upgraded to handle those side scenarios.

main.py is now a single-click single source of execution. It is now a matter of "python main.py" to update the dashboard.

Future supplies data is now supplied with info about who is responsible for the supply. This data will be used to display it on the forecast chart.